python-api/requirements.txt
flask
SQLAlchemy
PyMySQL
mysql-connector-python
bcrypt
flask_session
PyJWT
watchdog    